debugging push

+++ src/BitBoard.java
@@ -17,43 +17,43 @@ public class BitBoard {
         return bit^mask;
     }
 
-    public boolean checkHor( int unmaskedPosition){
-        int m = unmaskedPosition & (unmaskedPosition >> 7);
-        int f = m & (m >> 14);   
-        if(Integer.valueOf(f) > 0){
+    public boolean checkHor( long unmaskedPosition){
+        long m = unmaskedPosition & (unmaskedPosition >> 7);
+        long f = m & (m >> 14);   
+        if(Long.valueOf(f) > 0){
             return true;
         } 
         return false;
     } 
 
-    public boolean checkUpLeftDiag(int unmaskedPosition){
-        int m = unmaskedPosition & (unmaskedPosition >> 6);
-        int f = m & (m >> 12);   
-        if(Integer.valueOf(f) > 0){
+    public boolean checkUpLeftDiag(long unmaskedPosition){
+        long m = unmaskedPosition & (unmaskedPosition >> 6);
+        long f = m & (m >> 12);   
+        if(Long.valueOf(f) > 0){
             return true;
         } 
         return false;
     } 
 
-    public boolean checkUpRight(int unmaskedPosition){
-        int m = unmaskedPosition & (unmaskedPosition >> 8);
-        int f = m & (m >> 16);   
-        if(Integer.valueOf(f) > 0){
+    public boolean checkUpRight(long unmaskedPosition){
+        long m = unmaskedPosition & (unmaskedPosition >> 8);
+        long f = m & (m >> 16);   
+        if(Long.valueOf(f) > 0){
             return true;
         } 
         return false;
     } 
 
-    public boolean checkVert(int unmaskedPosition){
-        int m = unmaskedPosition & (unmaskedPosition >> 1);
-        int f = m & (m >> 2);   
-        if(Integer.valueOf(f) > 0){
+    public boolean checkVert(long unmaskedPosition){
+        long m = unmaskedPosition & (unmaskedPosition >> 1);
+        long f = m & (m >> 2);   
+        if(Long.valueOf(f) > 0){
             return true;
         } 
         return false;
     } 
 
-    public boolean checkWin(int unmaskedPosition){
+    public boolean checkWin(long unmaskedPosition){
         if(checkHor(unmaskedPosition) || checkUpLeftDiag(unmaskedPosition)||checkUpRight(unmaskedPosition)||checkVert(unmaskedPosition)){
             return true;
         }
